Edit deduplicated finding dialog own component
