+++ src/components/sections/judge/JudgeDeduplicatedFinding.tsx
@@ -1,16 +1,12 @@
 'use client'
 
 import Link from 'next/link'
-import {ArrowLeft, ArrowRight, Edit, Trash2, VerifiedIcon} from 'lucide-react'
+import {ArrowLeft, ArrowRight, Trash2, VerifiedIcon} from 'lucide-react'
 import {notFound} from 'next/navigation'
-import {useState} from 'react'
-import {useForm} from 'react-hook-form'
-import {zodResolver} from '@hookform/resolvers/zod'
-import {z} from 'zod'
 
 import JudgeAddDeduplicatedFinding from './JudgeAddDeduplicatedFinding'
 import FindingSeverityBadge from '../finding/FindingSeverityBadge'
-import FindingSeverityButtonSelect from '../finding/NewFindingForm/FindingSeverityButtonSelect'
+import JudgeEditDeduplicatedFindingFormDialog from './JudgeEditDeduplicatedFindingFormDialog'
 
 import {Button} from '@/components/ui/Button'
 import {PATHS} from '@/lib/utils/common/paths'
@@ -23,39 +19,14 @@ import {useSearchParamsEnumState} from '@/lib/hooks/useSearchParamsState'
 import {JudgeDeduplicatedFindingTab} from '@/server/db/models'
 import {useRemoveDededuplicatedFinding} from '@/lib/queries/deduplicatedFinding/removeDeduplicatedFinding'
 import {toast} from '@/components/ui/Toast'
-import {
-  DialogRoot,
-  DialogHeader,
-  DialogTrigger,
-  DialogContent,
-  DialogTitle,
-  DialogDescription,
-} from '@/components/ui/Dialog'
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '@/components/ui/Form'
-import {Input} from '@/components/ui/Input'
-import {editDeduplicatedFindingSchema} from '@/server/utils/validations/schemas'
-import {useEditDeduplicatedFinding} from '@/lib/queries/deduplicatedFinding/editDeduplicatedFinding'
-import Textarea from '@/components/ui/Textarea'
 
 type JudgeDeduplicatedFindingProps = {
   deduplicatedFindingId: string
 }
 
-type FormValues = z.infer<typeof editDeduplicatedFindingSchema>
-
 const JudgeDeduplicatedFinding = ({
   deduplicatedFindingId,
 }: JudgeDeduplicatedFindingProps) => {
-  const [isOpenEditDeduplicatedFinding, setIsOpenEditDeduplicatedFinding] =
-    useState(false)
-
   const [tab, {setValue: setTabValue}] = useSearchParamsEnumState(
     'tab',
     JudgeDeduplicatedFindingTab,
@@ -80,42 +51,6 @@ const JudgeDeduplicatedFinding = ({
             description:
               'Finding has been removed from this deduplicated finding.',
           })
-        },
-      },
-    )
-  }
-
-  const {mutate: editDeduplicatedFindingMutate} = useEditDeduplicatedFinding()
-
-  const form = useForm<FormValues>({
-    resolver: zodResolver(editDeduplicatedFindingSchema),
-    defaultValues: {
-      deduplicatedFindingId,
-      description: deduplicatedFinding?.description,
-      severity: deduplicatedFinding?.severity,
-      title: deduplicatedFinding?.title,
-    },
-  })
-
-  const editDeduplicatedFinding = ({
-    description,
-    severity,
-    title,
-  }: FormValues) => {
-    editDeduplicatedFindingMutate(
-      {
-        deduplicatedFindingId,
-        description,
-        severity,
-        title,
-      },
-      {
-        onSuccess: () => {
-          toast({
-            title: 'Success',
-            description: 'Deduplicated finding has been updated.',
-          })
-          setIsOpenEditDeduplicatedFinding(false)
         },
       },
     )
@@ -156,78 +91,14 @@ const JudgeDeduplicatedFinding = ({
           <div className="flex flex-col gap-3">
             <div className="flex items-center gap-4">
               <h1 className="text-headlineS">{deduplicatedFinding.title}</h1>
-              <DialogRoot
-                open={isOpenEditDeduplicatedFinding}
-                onOpenChange={setIsOpenEditDeduplicatedFinding}>
-                <DialogTrigger>
-                  <Button variant="ghost" className="flex gap-3">
-                    <span className="uppercase">
-                      <Edit />
-                    </span>
-                  </Button>
-                </DialogTrigger>
-
-                <DialogContent className="border-0 bg-grey-90">
-                  <Form {...form} onSubmit={editDeduplicatedFinding}>
-                    <DialogHeader>
-                      <DialogTitle className="text-titleM uppercase">
-                        Edit Deduplicated Finding
-                      </DialogTitle>
-                      <DialogDescription className="text-bodyM text-white">
-                        Edit the deduplicated finding details
-                      </DialogDescription>
-                    </DialogHeader>
-
-                    <div className="flex flex-col gap-8">
-                      <FormField
-                        control={form.control}
-                        name="title"
-                        render={({field}) => (
-                          <FormItem>
-                            <FormLabel>Title</FormLabel>
-                            <FormControl>
-                              <Input {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="severity"
-                        render={({field: {ref, ...field}}) => (
-                          <FormItem>
-                            <FormLabel>Choose the severity level</FormLabel>
-                            <FormControl>
-                              <FindingSeverityButtonSelect {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="description"
-                        render={({field: {ref, ...field}}) => (
-                          <FormItem>
-                            <FormLabel>Description</FormLabel>
-                            <FormControl>
-                              <Textarea {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-
-                    <div className="flex w-full justify-end">
-                      <Button type="submit">Save</Button>
-                    </div>
-                  </Form>
-                </DialogContent>
-              </DialogRoot>
+              <JudgeEditDeduplicatedFindingFormDialog
+                deduplicatedFindingId={deduplicatedFindingId}
+                defaultValues={{
+                  title: deduplicatedFinding.title,
+                  description: deduplicatedFinding.description,
+                  severity: deduplicatedFinding.severity,
+                }}
+              />
             </div>
             <div>
               <FindingSeverityBadge severity={deduplicatedFinding.severity} />
